backend/src/requirements.txt
numpy==1.18.2
Pillow==7.1.1
tqdm==4.46.0
tensorflow==2.2.0
